[js] Put open and close logic into separate functions
Previously the open/close state of pageguide was stored in two places:
in the object property 'open' on the page guide, and on the body element
of the page, depending on the presence of the 'tlypageguide-open' class.
Now it's just stored in the body tag.

This this state should probably be encapsulated in the pageguide object.